Decompose Review.tsx into focused components
Review.tsx (4,323 lines) is split into a review/ component directory:
DiffView, CommentThread, InlineCommentForm, PRHeader, ReviewDiscussion,
FileIndex, AddCommentModal, ReviewSubmitModal, PluginsPanel,
OutdatedCommentsPanel, MobileReviewBar, CopyUrlButton, plus shared
types/helpers (types.ts) and the Prism diff theme builder (diff_theme.ts).
Review.tsx (now 1,427 lines) keeps state, RPC handlers, the toolbar, and
the dock/split-panel layout.

Along the way this removes two long-standing duplications:
- The inline comment thread and comment editor were copy-pasted between
  the file-header and code-line branches of renderDiff; both now render
  the shared CommentThread / InlineCommentForm components.
- diff_utils.ts held a stale fork of the diff parser while Review.tsx
  carried a newer inline copy. The newer implementation (line numbers,
  hunk context, rename/quoted-filename handling, expanded-context lines)
  is now the single exported parseDiff, still covered by the existing
  regression tests, alongside the extracted test-file ordering helpers.

Also fixes a latent highlight mismatch: syntax-highlighted lines were
keyed by unsorted parse order but looked up in tests-last sorted order,
so diffs containing test files could highlight the wrong rows.
