+++ bun_client/frontend/src/components/review/types.ts
@@ -1,0 +1,193 @@
+import { colors } from '../../design';
+
+// Above this many changed lines in a single file, the diff renders that file
+// without per-line syntax highlighting to avoid thousands of Prism instances.
+export const MAX_HIGHLIGHT_LINES_PER_FILE = 2000;
+
+// Stable slug for in-page anchor IDs from file paths.
+export const slugify = (s: string): string =>
+    s.replace(/[^a-zA-Z0-9_-]+/g, '-').replace(/^-+|-+$/g, '');
+
+// Strip HTML comments from text (e.g., <!-- comment -->)
+export const stripHtmlComments = (text: string): string => {
+    return text.replace(/<!--[\s\S]*?-->/g, '');
+};
+
+export interface Comment {
+    id: string;
+    author: string;
+    body: string;
+    path: string;
+    position: string;
+    in_reply_to: number;
+    created_at: string;
+    outdated: boolean;
+    diff_hunk: string;
+}
+
+export interface ReviewData {
+    id: number;
+    user: string;
+    body: string;
+    state: string;
+    submitted_at: string;
+    html_url: string;
+}
+
+export interface PluginResult {
+    result: string;
+    status: string;
+}
+
+export interface PRMetadata {
+    number: number;
+    title: string;
+    author: string;
+    base_ref: string;
+    head_ref: string;
+    state: string;
+    milestone: string;
+    labels: string[];
+    assignees: string[];
+    reviewers: string[]; // Requested individual reviewers
+    requested_teams: string[]; // Requested team reviewers
+    approved_by: string[]; // Users who approved
+    changes_requested_by: string[]; // Users who requested changes
+    commented_by: string[]; // Users who commented
+    draft: boolean;
+    ci_status: string;
+    ci_failures: string[];
+    body: string;
+    url: string;
+    repo_path: string;
+    worktree_path: string;
+}
+
+export interface PRResponse {
+    content: string;
+    diff: string;
+    comments: Comment[];
+    outdated_comments: Comment[];
+    reviews: ReviewData[];
+    metadata: PRMetadata;
+    feedback: string;
+    // Only set by SyncPR: true when the sync pulled in a new head SHA or new comments.
+    updated?: boolean;
+}
+
+// Map file extensions to Prism language identifiers
+export const getLanguageFromFilename = (filename: string): string => {
+    const ext = filename.split('.').pop()?.toLowerCase() || '';
+    const languageMap: Record<string, string> = {
+        // JavaScript/TypeScript
+        js: 'javascript',
+        jsx: 'jsx',
+        ts: 'typescript',
+        tsx: 'tsx',
+        mjs: 'javascript',
+        cjs: 'javascript',
+        // Web
+        html: 'html',
+        htm: 'html',
+        css: 'css',
+        scss: 'scss',
+        sass: 'sass',
+        less: 'less',
+        json: 'json',
+        xml: 'xml',
+        svg: 'xml',
+        // Backend
+        py: 'python',
+        rb: 'ruby',
+        go: 'go',
+        rs: 'rust',
+        java: 'java',
+        kt: 'kotlin',
+        scala: 'scala',
+        php: 'php',
+        c: 'c',
+        h: 'c',
+        cpp: 'cpp',
+        cc: 'cpp',
+        cxx: 'cpp',
+        hpp: 'cpp',
+        cs: 'csharp',
+        swift: 'swift',
+        // Shell/Config
+        sh: 'bash',
+        bash: 'bash',
+        zsh: 'bash',
+        fish: 'bash',
+        ps1: 'powershell',
+        yaml: 'yaml',
+        yml: 'yaml',
+        toml: 'toml',
+        ini: 'ini',
+        conf: 'ini',
+        // Data/Docs
+        md: 'markdown',
+        mdx: 'markdown',
+        sql: 'sql',
+        graphql: 'graphql',
+        gql: 'graphql',
+        // Other
+        dockerfile: 'docker',
+        makefile: 'makefile',
+        lua: 'lua',
+        r: 'r',
+        dart: 'dart',
+        ex: 'elixir',
+        exs: 'elixir',
+        erl: 'erlang',
+        hrl: 'erlang',
+        clj: 'clojure',
+        vim: 'vim',
+        el: 'lisp',
+        lisp: 'lisp',
+        hs: 'haskell',
+        ml: 'ocaml',
+        proto: 'protobuf',
+    };
+
+    // Handle special filenames
+    const basename = filename.split('/').pop()?.toLowerCase() || '';
+    if (basename === 'dockerfile') return 'docker';
+    if (basename === 'makefile' || basename === 'gnumakefile') return 'makefile';
+    if (basename.endsWith('.d.ts')) return 'typescript';
+
+    return languageMap[ext] || 'text';
+};
+
+// Get icon and color for file status
+export const getFileStatusInfo = (status?: 'modified' | 'new' | 'deleted' | 'renamed') => {
+    switch (status) {
+        case 'new':
+            return {
+                icon: '+',
+                label: 'new file',
+                color: colors.success,
+                bg: colors.bgSuccessDim,
+            };
+        case 'deleted':
+            return {
+                icon: '−',
+                label: 'deleted',
+                color: colors.danger,
+                bg: colors.bgDangerDim,
+            };
+        case 'renamed':
+            return {
+                icon: '→',
+                label: 'renamed',
+                color: colors.warning,
+                bg: colors.bgWarningDim,
+            };
+        default:
+            return {
+                icon: '●',
+                label: 'modified',
+                color: colors.accent,
+                bg: colors.bgInfoDimStrong,
+            };
+    }
+};
